+++ src/actions/index.js
@@ -1,8 +1,6 @@
 import 'isomorphic-fetch';
-// import cookie from 'react-cookie';
 
 // STATUS
-// Dev-only actions for backend testing
 
 export const GET_STATUS = 'GET_STATUS';
 export const getStatus = () => ({
@@ -18,13 +16,9 @@ export const getStatusSuccess = message => ({
 export const fetchStatus = () => (dispatch) => {
   dispatch(getStatus());
 
-  fetch('http://localhost:8080/', {
-    // headers: {
-    //   Authorization: `Bearer ${cookie.load('accessToken')}`,
-    // },
-  })
+  fetch('http://localhost:8080/')
+
   .then((res) => {
-    console.info('get status');
     if (!res.ok) {
       const error = new Error(res.statusText);
       error.response = res;
@@ -32,103 +26,10 @@ export const fetchStatus = () => (dispatch) => {
     }
     return res;
   })
+
   .then(res => res.json())
+
   .then(data => dispatch(getStatusSuccess(data.message)))
 
-  // TODO fetch error action
   .catch(console.error);
-};
-
-// PHOTOS
-
-export const GET_PHOTOS = 'GET_PHOTOS';
-export const getPhotos = () => ({
-  type: GET_PHOTOS,
-});
-
-export const GET_PHOTOS_SUCCESS = 'GET_PHOTOS_SUCCESS';
-export const getPhotosSuccess = photos => ({
-  type: GET_PHOTOS_SUCCESS,
-  photos,
-});
-
-const mockPhotos = [
-  {
-    userId: 'Jamie',
-    url: 'http://lorempixel.com/800/600/people/',
-    date: new Date(),
-    tags: ['Jamie', 'Grandma', 'Spot'],
-  },
-  {
-    userId: 'Alex',
-    url: 'http://lorempixel.com/600/800/people/',
-    date: new Date(),
-    tags: ['Alex', 'Grandma', 'Spot'],
-  },
-  {
-    userId: 'Ben',
-    url: 'http://lorempixel.com/1000/800/city/',
-    date: new Date(),
-    tags: ['Jamie', 'Ben', 'Spot'],
-  },
-  {
-    userId: 'Grandma',
-    url: 'http://lorempixel.com/600/800/sports/',
-    date: new Date(),
-    tags: ['Jimmy', 'Grandma', 'Spot'],
-  },
-];
-
-export const fetchPhotos = () => (dispatch) => {
-  dispatch(getPhotos());
-
-  fetch('http://localhost:8080/', {
-    // headers: {
-    //   Authorization: `Bearer ${cookie.load('accessToken')}`,
-    // },
-  })
-  .then((res) => {
-    console.log('get photo');
-    if (!res.ok) {
-      const error = new Error(res.statusText);
-      error.response = res;
-      throw error;
-    }
-    return res;
-  })
-  .then(res => res.json())
-  .then(() => dispatch(getPhotosSuccess(mockPhotos)))
-
-  // TODO fetch error action
-  .catch(console.error);
-};
-
-// PEOPLE
-
-const mockPeople = {
-  Jamie: {
-    avatar: '...url...',
-    fullname: 'Jamie Davella',
-  },
-  Ben: {
-    avatar: '...url...',
-    fullname: 'Ben Fletcher',
-  },
-};
-
-export const GET_PEOPLE = 'GET_PEOPLE';
-export const getPeople = () => ({
-  type: GET_PEOPLE,
-});
-
-export const GET_PEOPLE_SUCCESS = 'GET_PEOPLE_SUCCESS';
-export const getPeopleSuccess = people => ({
-  type: GET_PEOPLE_SUCCESS,
-  people,
-});
-
-export const fetchPeople = () => (dispatch) => {
-  dispatch(getPeople());
-
-  dispatch(getPeopleSuccess(mockPeople));
 };
